requirements.txt
pandas
numpy
sqlalchemy
mysql-connector-python
faker
python-dotenv
mlflow